requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alpaca-trade-api==3.2.0
alpha_vantage==3.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
confluent-kafka==2.6.1
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
influxdb-client==1.48.0
jiter==0.8.0
joblib==1.4.2
kafka-python==2.0.2
msgpack==1.0.3
multidict==6.1.0
numpy==2.1.3
openai==0.28.0
packaging==24.2
pandas==2.2.3
plyer==2.1.0
propcache==0.2.0
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
reactivex==4.0.4
redis==5.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.18.0
pyswarms==1.3.0
# pygmo==2.19.0
swarmops==1.0.0
ants==0.0.7

